Completion Request should return CompletionList
As discussed in
https://github.com/Microsoft/language-server-protocol/issues/39
statically typed languages like Java do not support mixed return types,
which in the case of Completion Request is `CompletionItem[] |
CompletionList`.

Thus Typefox - the Java client of the language server protocol - accepts
only `CompletionList`, which contains `CompletionItem[]`.

In order to work with Eclipse Che and other IDEs using the Typefox
client, the Completion Request handler in the PHP language server should
return `CompletionList` instead of `CompletionItems[]`.