CompletionItems should report textEdit instead of insertText
The `textEdit` property of `CompletionItems contains more details than
`insertText`. In addition to what text should be inserted, it tells what
part of the document should be replaced.

Using `textEdit` is better for unified behavior with different clients.
For example VSCode works pretty well with just `insertText` because it
has additional logic on the client side to compute what range should be
replaced in the editor. Eclipse Che does not implement such logic and
the user experience with code completion is not that good (with
`insertText`). Using `textEdit` brings equal behavior in both clients.

This commit is an early attempt to compute adequate `textEdit` ranges.
It should be improved with:
- better determination of the node-at-cursor-position to be replaced
- better fallback calculation of the range when no node can be
determined by the parser at the cursor position